비밀번호 변경 커밋
확인했습니다

+++ src/main/java/com/team1/stelling/domain/vo/UserVO.java
@@ -31,13 +31,13 @@ public class UserVO {
     @Column(name = "USER_EMAIL")
     private String userEmail;   // 이메일
     @Column(name = "USER_GENDER")
-    private Long userGender;  // 성별
+    private int userGender;  // 성별
     @Column(name = "USER_PHONENUM")
     private String userPhoneNum;  //전화번호
     @Column(name = "USER_STATUS")
-    private Long userStatus;      // 회원 상태 0 - 일반회원 , 1 - 관리자 , 2- 휴면(탈퇴)
+    private int userStatus;      // 회원 상태 0 - 일반회원 , 1 - 관리자 , 2- 휴면(탈퇴)
     @Column(name = "USER_COINBALANCE")
-    private Long userCoinBalance;  // 코인 잔액
+    private int userCoinBalance;  // 코인 잔액
     @Column(name = "USER_FILEPATH")
     private String userFilePath;  // 파일 경로
     @Column(name = "USER_UUID")
@@ -50,19 +50,20 @@ public class UserVO {
     public void updateUserPw(String userPw) { this.userPw = userPw; }
     public void updateUserNickName(String userNickName) { this.userNickName = userNickName; }
     public void updateUserEmail(String userEmail) { this.userEmail = userEmail; }
-    public void updateUserGender(Long userGender) { this.userGender = userGender; }
+    public void updateUserGender(int userGender) { this.userGender = userGender; }
     public void updateUserPhoneNum(String userPhoneNum) { this.userPhoneNum = userPhoneNum; }
-    public void updateUserStatus(Long userStatus) { this.userStatus = userStatus; }
-    public void updateUserCoinBalance(Long userCoinBalance) { this.userCoinBalance = userCoinBalance; }
+    public void updateUserStatus(int userStatus) { this.userStatus = userStatus; }
+    public void updateUserCoinBalance(int userCoinBalance) { this.userCoinBalance = userCoinBalance; }
     public void updateUserAccessToken(String userAccessToken) { this.userAccessToken = userAccessToken; }
     public void updateUserUuid(String userUuid){this.userUuid = userUuid;}
     public void updateFilePate(String userFilePath){this.userFilePath = userFilePath;}
     public void updateFileName(String userFileName){this.userFileName = userFileName;}
 
     @Builder
-    public UserVO(Long userNumber, String userId, String userPw, String userNickName, String userEmail,
-                  Long userGender, String userPhoneNum, Long userStatus, Long userCoinBalance, String userFilePath,
-                  String userUuid, String userFileName, String userAccessToken)
+    public UserVO(Long userNumber, String userId, String userPw, String userNickName,
+                  String userEmail, int userGender, String userPhoneNum,
+                  int userStatus, int userCoinBalance, String userFilePath, String userUuid,
+                  String userFileName, String userAccessToken)
     {
         this.userNumber = userNumber;
         this.userId = userId;
